fix: renamed auto_routes to auth
